feat: Do not return `null` as analyticsUrl for CI=true
BREAKING CHANGE: Expose analytics server URL unconditionally (and not just
in non-CI environments) It is breaking as it might impact how analytics
reporting works in older version of the Framework and might pollute data
with unexpected events coming from CI deployments. Additonally, it will
cause notifications to be triggered in CI environments as well.